+++ NEWS.rst
@@ -33,7 +33,7 @@ Changes in 3.6 (released 12/31/2008)
 *	Fixed a bug in :meth:`ll.make.CacheAction.get`: The action must return real
 	data when called with ``bigbang`` as the timestamp.
 
-*	:class:`ll.make.UL4RenderAction`` has been fixed.
+*	:class:`ll.make.UL4RenderAction` has been fixed.
 
 
 Changes in 3.5 (released 12/05/2008)